テストコード実装完了

+++ resources/js/components/hot/Hot.vue
@@ -546,10 +546,6 @@ export default {
       // 無限スクロール用
       postsLoading: false,
       loadMorePosts: true,
-      // itemLoading: false,
-      // loadMore: true,
-      // page: 1,
-      // isLastPage: false,
       // 投稿
       posts: [
         // {
@@ -631,8 +627,6 @@ export default {
       // コメントの無限スクロール用
       commentsLoading: false,
       loadCommentsMore: true,
-      // commentsPage: 1,
-      // isCommentsLastPage: false,
       // 新規コメント、返信
       newComment: {},
       newReply: {},
@@ -643,6 +637,7 @@ export default {
       // 削除の確認のモーダル
       deletePostModalOpened: false,
       deletePostIndex: null,
+      deleteProssesing: false,
       // 投稿編集モーダル
       modalPostEditShow: false,
       editPostIndex: null,
@@ -766,12 +761,16 @@ export default {
     },
     // 投稿の削除
     deletePost() {
+      if (this.deleteProssesing) return;
+      this.deleteProssesing = true;
       axios.post('/api/post/delete/' + this.posts[this.deletePostIndex].id)
         .then(() => {
           this.posts.splice(this.deletePostIndex, 1);
           this.deletePostModalClose();
-        }).catch(() => {
-          return;
+          this.deleteProssesing = false;
+        }).catch((error) => {
+          console.log(error);
+          this.deleteProssesing = false;
         });
     },
     // 投稿編集モーダルを開く
@@ -826,7 +825,7 @@ export default {
       this.editErrors = [];
       this.nextNewImageId = -1;
       this.editHeight = '20px';
-      if (!this.edited) {
+      if (!this.editProcessing) {
         this.closePostMenu(this.editPostIndex);
       }
       this.modalPostEditShow = false;
@@ -903,9 +902,7 @@ export default {
       axios.post('/api/post/edit', data)
         .then((res) => {
           this.posts.splice(this.editPostIndex, 1, res.data.post);
-          this.edited = true;
           this.editPostModalClose();
-          this.edited = false;
           this.editProcessing = false;
         }).catch((error) => {
           this.editErrors = error.response.data.errors;
@@ -1039,7 +1036,10 @@ export default {
       axios.get('/api/donmai/users/' + this.modalPostId + '?page=' + this.donmaiPage)
         .then((res) => {
           console.log(res.data);
-          this.modalDonmaiUsers.push(...res.data.data);
+          const users = res.data.data.map((obj) => {
+            return obj.user;
+          });
+          this.modalDonmaiUsers.push(...users);
           this.donmaiLoading = false;
           if (this.donmaiPage === res.data.last_page) {
             this.isLastDonmaiPage = true;
@@ -1085,9 +1085,10 @@ export default {
       const img_height = img.height;
       if (img_height >= img_width) {
         this.heightIsBigger = true;
-        document.querySelector('.overlay-image-image').addEventListener('load', () => {
-          this.tatenagaImageWidth = document.querySelector('.overlay-image-image').clientWidth;
-          // console.log(this.tatenagaImageWidth);
+        this.$nextTick(function() {
+            this.tatenagaImageWidth = document.querySelector('.overlay-image-image').clientWidth;
+            this.handleResize();
+            // console.log(this.tatenagaImageWidth);
         });
       }
     },
